renamed node to heat-capacitor
