+++ docs/images/heat-capacitor-temperatureVsPrice.png
