+++ docs/images/node-ps-strategy-heat-capacitor-simple-flow-example.png
