+++ docs/nodes/ps-strategy-heat-capacitor.md
