+++ package.json
@@ -25,7 +25,7 @@
       "ps-receive-price": "src/receive-price.js",
       "ps-strategy-best-save": "src/strategy-best-save.js",
       "ps-strategy-lowest-price": "src/strategy-lowest-price.js",
-      "ps-strategy-temperature-manipulation": "src/strategy-temperature-manipulation.js",
+      "ps-strategy-heat-capacitor": "src/strategy-heat-capacitor.js",
       "ps-elvia-config": "src/elvia/elvia-config.js",
       "ps-elvia-tariff-types": "src/elvia/elvia-tariff-types.js",
       "ps-elvia-tariff": "src/elvia/elvia-tariff.js",
