+++ src/strategy-heat-capacitor.html
@@ -1,5 +1,5 @@
 <script type="text/javascript">
-    RED.nodes.registerType('ps-strategy-temperature-manipulation',{
+    RED.nodes.registerType('ps-strategy-heat-capacitor',{
         category: "Power Saver",
         color: "#FFCC66",
         defaults: {
@@ -20,7 +20,7 @@
     });
 </script>
 
-<script type="text/html" data-template-name="ps-strategy-temperature-manipulation">
+<script type="text/html" data-template-name="ps-strategy-heat-capacitor">
     <div class="form-row">
         <label for="node-input-name"><i class="icon-tag"></i> Name</label>
         <input type="text" id="node-input-name" placeholder="Name">
